requirements.txt
selenium>=4.0
webdriver-manager>=4.0
python-dotenv>=1.0
cryptography>=42.0